ES query: Muevo código de formateo de respuesta a otro módulo
    refs:#236
