+++ series_tiempo_ar_api/apps/api/query/es_query/response_formatter.py
@@ -1,0 +1,86 @@
+#! coding: utf-8
+from iso8601 import iso8601
+
+from series_tiempo_ar_api.apps.api.helpers import get_relative_delta
+from series_tiempo_ar_api.apps.api.query import constants
+
+
+class ResponseFormatter(object):
+
+    def __init__(self, series, responses, args, periodicity):
+        self.series = series
+        self.responses = responses
+        self.data_dict = {}
+        self.args = args
+        self.periodicity = periodicity
+
+    def format_response(self):
+        """Procesa la respuesta recibida de Elasticsearch, la guarda en
+        el diccionario data_dict con el siguiente formato
+        self.data_dict = {
+          "1990-01-01": { "serie_1": valor1, "serie_2": valor2, ... },
+          "1990-02-01": { "serie_1": valor1, "serie_2": valor2, ... }
+        }
+        Luego el diccionario es pasado a la lista de listas final
+        self.data para conformar la respuesta esperada de lista de listas
+        """
+        final_data = []
+        for i, response in enumerate(self.responses):
+            rep_mode = self.series[i].rep_mode
+
+            for hit in response:
+                data = hit[rep_mode] if rep_mode in hit else None
+                timestamp_dict = self.data_dict.setdefault(hit.timestamp, {})
+                timestamp_dict[self._data_dict_series_key(self.series[i])] = data
+
+        if not self.data_dict:  # No hay datos
+            return []
+
+        self._make_date_index_continuous(min(self.data_dict.keys()),
+                                         max(self.data_dict.keys()))
+
+        # Ordeno las timestamp según si el sort es asc o desc usando función de comparación
+        def cmp_func(one, other):
+            if one == other:
+                return 0
+
+            if self.args[constants.PARAM_SORT] == constants.SORT_ASCENDING:
+                return -1 if one < other else 1
+            else:
+                return 1 if one < other else -1
+
+        for timestamp in sorted(self.data_dict.keys(), cmp=cmp_func):
+            row = [timestamp]
+
+            for series in self.series:
+                row.append(self.data_dict[timestamp].get(self._data_dict_series_key(series)))
+
+            final_data.append(row)
+
+        return final_data
+
+    @staticmethod
+    def _data_dict_series_key(series):
+        """Key única para identificar la serie pedida en el data_dict armado. Evita
+        que se pisen series en queries que piden la misma serie con distintos rep modes
+        o aggs (ver issue #243)
+        """
+        return series.series_id + series.rep_mode + series.collapse_agg
+
+    def _make_date_index_continuous(self, start_date, end_date):
+        """Hace el índice de tiempo de los resultados continuo (según
+        el intervalo de resultados), sin saltos, entre start_date y end_date.
+        Esto implica llenar el diccionario self.data_dict con claves de los
+        timestamp faltantes para asegurar la continuidad
+        """
+
+        # Si no hay datos cargados no hay nada que hacer
+        if not len(self.data_dict):
+            return
+
+        current_date = iso8601.parse_date(start_date)
+        end_date = iso8601.parse_date(end_date)
+
+        while current_date < end_date:
+            current_date += get_relative_delta(self.periodicity)
+            self.data_dict.setdefault(unicode(current_date.date()), {})
